requirements.txt
alchemical==0.5.1
alembic==1.7.7
apifairy==1.3.0
apispec==5.1.1
blinker==1.4
certifi==2022.12.7
click==8.1.0
colorama==0.4.6
Faker==13.3.4
Flask==2.1.0
Flask-Authorize==0.2.6
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
packaging==21.3
Pillow==9.2.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
sentry-sdk==1.11.1
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.26.13
webargs==8.1.0
Werkzeug==2.1.0
